requirements.txt
bcrypt==3.2.0
cffi==1.15.0
click==8.0.4
cryptography==36.0.1
Flask==2.0.3
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
python-dotenv==0.19.2
six==1.16.0
SQLAlchemy==1.4.31
Werkzeug==2.0.3
